backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# FHIR
fhir.resources==7.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-yasg==1.21.7

# Environment
python-decouple==3.8

# WSGI Server
gunicorn==21.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==22.6.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# HTTP Client for healthcheck
requests==2.31.0

# Caching
redis==4.6.0
django-redis==5.4.0

# Message Queue & Async Tasks
celery[redis]==5.3.4
kombu==5.3.4
amqp==5.2.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Report Generation
reportlab==4.0.8
pandas==2.1.4
openpyxl==3.1.2
